Extract EditableMonitor class into its own file
